requirements.txt
alembic==1.7.7
APScheduler==3.9.1
beautifulsoup4==4.11.1
blinker==1.4
bs4==0.0.1
cffi==1.15.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.8
Flask==1.1.4
Flask-APScheduler==1.12.3
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-pyjwt==0.1.9
Flask-PyMongo==2.3.0
Flask-pymysql==0.2.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.2.0
MarkupSafe==2.0.1
numpy==1.22.3
pandas==1.4.2
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pymongo==4.1.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.17.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.35
SQLAlchemy-Utils==0.38.2
tzdata==2022.1
tzlocal==4.2
Werkzeug==1.0.1
zipp==3.8.0
